ft_printf done, all mandatory tests pass

+++ ft_printf.h
@@ -3,10 +3,10 @@
 /*                                                        :::      ::::::::   */
 /*   ft_printf.h                                        :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: noscrub <noscrub@student.42.fr>            +#+  +:+       +#+        */
+/*   By: ncontin <ncontin@student.42.fr>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/10/21 16:36:18 by ncontin           #+#    #+#             */
-/*   Updated: 2024/10/23 16:31:24 by noscrub          ###   ########.fr       */
+/*   Updated: 2024/10/23 18:13:47 by ncontin          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -15,7 +15,6 @@
 
 # include <stdarg.h>
 # include <stdio.h>
-# include <stdlib.h>
 # include <unistd.h>
 
 int	ft_putchar(char c);
@@ -23,7 +22,7 @@ int	ft_putstr(char *str);
 int	ft_putnbr(int n);
 int	ft_putnbr_uns(unsigned int n);
 int	ft_printf(const char *format, ...);
-int	ft_puthex(unsigned long int n, char hex_case);
+int	ft_puthex(unsigned int n, char hex_case);
 int	ft_putptr(void *n);
 
 #endif
